+++ R/get_compressed_sdmx.R
@@ -25,17 +25,19 @@ get_compressed_sdmx<-function(url=NULL,verbose=FALSE,format="gz"){
     return(NULL)
   } else {
     if (tbc) {
+      if (verbose) message("get_compressed_sdmx - url:",url,"\nget_compressed_sdmx - format:",format)
+
       if (format=="zip"){ # zip format
         # download file 
         temp<-tempfile()  
         if (verbose) {
           tryCatch({utils::download.file(url,temp,dmethod)},
                    error = function(e) {
-                     message("\nget_compressed_sdmx - Error by the download of the SDMX file:",'\n',paste(unlist(e),collapse="\n"))
+                     message("get_compressed_sdmx - Error by the download of the SDMX file:",'\n',paste(unlist(e),collapse="\n"))
                      tbc<-FALSE
                    },
                    warning = function(w) {
-                     message("\nget_compressed_sdmx - Warning during the download of the SDMX file:",'\n',paste(unlist(w),collapse="\n"))
+                     message("get_compressed_sdmx - Warning during the download of the SDMX file:",'\n',paste(unlist(w),collapse="\n"))
                      tbc<-FALSE
                    })
         } else {
@@ -55,10 +57,10 @@ get_compressed_sdmx<-function(url=NULL,verbose=FALSE,format="gz"){
         if (verbose) {
           tryCatch({xml_fajl<-utils::unzip(temp,paste0(fajl,".xml"),exdir=tmpdir)},
                    error = function(e) {
-                     message("\nget_compressed_sdmx - Error during the unzip of the SDMX file:",'\n',paste(unlist(e),collapse="\n"))
+                     message("get_compressed_sdmx - Error during the unzip of the SDMX file:",'\n',paste(unlist(e),collapse="\n"))
                    },
                    warning = function(w) {
-                     message("\nget_compressed_sdmx - Warning by the unzip of the SDMX file:",'\n',paste(unlist(w),collapse="\n"))
+                     message("get_compressed_sdmx - Warning by the unzip of the SDMX file:",'\n',paste(unlist(w),collapse="\n"))
                    })
         } else {
           tryCatch({xml_fajl<-utils::unzip(temp,paste0(fajl,".xml"),exdir=tmpdir)},
@@ -70,23 +72,24 @@ get_compressed_sdmx<-function(url=NULL,verbose=FALSE,format="gz"){
           unlink(file.path(tmpdir,paste0(fajl,".xml")))
         } else if (format=="gz"){
         if (verbose) {
-          tryCatch({xml_fajl<-readLines(gzcon(url(url)),warn=verbose)},
+          tryCatch({xml<-xml2::read_xml(gzcon(url(url,open="rb")))},
                  error = function(e) {
-                   message("\nget_compressed_sdmx - Error during the unzip of the SDMX file:",'\n',paste(unlist(e),collapse="\n"))
+                   message("get_compressed_sdmx - Error during retrieval and extraction of the SDMX file:",'\n',paste(unlist(e),collapse="\n"))
                  },
                  warning = function(w) {
-                   message("\nget_compressed_sdmx - Warning by the unzip of the SDMX file:",'\n',paste(unlist(w),collapse="\n"))
+                   message("get_compressed_sdmx - Warning by the retrieval and extraction of the SDMX file:",'\n',paste(unlist(w),collapse="\n"))
                  })
         } else {
-          tryCatch({xml_fajl<-readLines(gzcon(url(url)),warn=verbose)},
+          tryCatch({xml<-xml2::read_xml(gzcon(url(url,open="rb")))},
                  error = function(e) {},
                  warning = function(w) {})
-        }  
-        if (!is.null(xml_fajl)){xml<-xml2::read_xml(readLines(xml_fajl,warn=FALSE))} else {xml<-NULL}
+        }
+        if (verbose) message("get_compressed_sdmx - xml NULL:",is.null(xml))  
+        # if (!is.null(xml_fajl)){xml<-xml2::read_xml(gzcon(url(url,open="rb")))} else {xml<-NULL}
         
 
       } else {
-        message("\nIncorrect compression format. The compression format should be either 'gz' or 'zip'.") 
+        message("Incorrect compression format. The compression format should be either 'gz' or 'zip'.") 
       }
     }  
   }